refactor: separate components in SignInScreen into individual files
This commit refactors the SignInScreen by separating the GoogleSignInButton and RememberFirebaseAuthLauncher components into their own individual files. This will test SonarCloud in a first time. It improves code organization and maintainability.
